fix: error in HarvestProcessor cleanup
Pre-declaring cleanup file attributes was causing None to be fed into os.path.exists()

+++ src/galv_harvester/harvest.py
@@ -468,7 +468,7 @@ class HarvestProcessor:
         for attribute in ["data_file_name", "png_file_name"]:
             if hasattr(self, attribute):
                 filename = getattr(self, attribute)
-                if os.path.exists(filename):
+                if filename is not None and os.path.exists(filename):
                     try:
                         if os.path.isdir(filename):
                             shutil.rmtree(filename)
